BugzID: 5866
JWebUnit upgrade. Includes all clients in tests now.




git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@21057 908b024a-c82b-635c-c293-f2b15b23277c

+++ test/org/zfin/sequence/reno/MultiRunTest.java
@@ -264,7 +264,7 @@ public class MultiRunTest {
             session.refresh(runCandidate2) ;
             assertEquals("runCandidate1 occurrs 1st",1,runCandidate1.getOccurrenceOrder()) ;
             assertEquals("runCandidate2 occurrs 2nd",2,runCandidate2.getOccurrenceOrder()) ;
-            assertEquals("run count should be 2",2,runCandidate1.getCandidate().getRunCount()) ;
+            assertEquals("run count should be 2",2,(int) runCandidate1.getCandidate().getRunCount()) ;
 
 
             NomenclatureCandidateController nomenclatureCandidateController = new NomenclatureCandidateController() ;
@@ -280,8 +280,8 @@ public class MultiRunTest {
             assertFalse("runCandidate2 is NOT done",runCandidate2.isDone()) ;
 
             session.refresh(runCandidate1.getCandidate()) ;
-            assertEquals("run count should be 2",2,runCandidate1.getCandidate().getRunCount()) ;
-            assertEquals("run count should be 2",2,runCandidate1.getCandidate().getRunCount()) ;
+            assertEquals("run count should be 2",2,(int) runCandidate1.getCandidate().getRunCount()) ;
+            assertEquals("run count should be 2",2,(int) runCandidate1.getCandidate().getRunCount()) ;
         }
         catch (Exception e) {
             java.lang.StackTraceElement[] elements = e.getStackTrace();
